requirements_whisper.txt
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
assemblyai==0.25.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
av==10.0.0
cached-property==1.5.2
certifi==2023.11.17
charset-normalizer==3.3.2
coloredlogs==15.0.1
ctranslate2==3.24.0
dataclasses-json==0.6.4
decorator==5.1.1
deepgram-sdk==3.2.5
distro==1.9.0
elevenlabs==0.2.27
exceptiongroup==1.2.0
executing==2.0.1
faster-whisper==0.10.0
ffmpeg-python==0.2.0
filelock==3.13.1
flatbuffers==23.5.26
frozenlist==1.4.1
fsspec==2023.12.2
future==0.18.3
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
huggingface-hub==0.20.3
humanfriendly==10.0
idna==3.6
ipython==8.20.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
MarkupSafe==2.1.4
marshmallow==3.21.1
matplotlib-inline==0.1.6
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
networkx==3.2.1
numpy==1.26.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
onnxruntime==1.16.3
openai==1.10.0
packaging==23.2
pandas==2.2.0
parso==0.8.3
pexpect==4.9.0
prompt-toolkit==3.0.43
protobuf==4.25.2
ptyprocess==0.7.0
pure-eval==0.2.2
PyAudio==0.2.14
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.17.2
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
safetensors==0.4.2
scikit-learn==1.4.0
scipy==1.12.0
six==1.16.0
skops==0.9.0
sniffio==1.3.0
stack-data==0.6.3
sympy==1.12
tabulate==0.9.0
threadpoolctl==3.2.0
tokenizers==0.15.1
torch==2.1.2
torchaudio==2.1.2
tqdm==4.66.1
traitlets==5.14.1
transformers==4.37.2
triton==2.1.0
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
verboselogs==1.7
wcwidth==0.2.13
websocket-client==1.7.0
websockets==12.0
-e git+https://github.com/RKelln/WhisperLive.git@d8ddb54fb885bb210ef79591eed62e6a33543860#egg=whisper_live
wtpsplit==1.3.0
yarl==1.9.4
